Refactor clases
