+++ app/src/main/java/com/example/android/vinylsappg21/viewmodels/ArtistViewModel.kt
@@ -1,10 +1,10 @@
-package com.example.android.vinylsappg21.ui.artists
+package com.example.android.vinylsappg21.viewmodels
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 
-class ArtistsViewModel : ViewModel() {
+class ArtistViewModel : ViewModel() {
 
     private val _text = MutableLiveData<String>().apply {
         value = "This is artists Fragment"
