+++ app/src/main/java/com/example/android/vinylsappg21/viewmodels/CollectorViewModel.kt
@@ -1,13 +1,13 @@
-package com.example.android.vinylsappg21.ui.albums
+package com.example.android.vinylsappg21.viewmodels
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 
-class AlbumsViewModeld : ViewModel() {
+class CollectorViewModel : ViewModel() {
 
     private val _text = MutableLiveData<String>().apply {
-        //value = "This is albums Fragment"
+        value = "This is collectors Fragment"
     }
     val text: LiveData<String> = _text
 }
