--- app/src/main/java/com/example/android/vinylsappg21/viewmodels/CommentViewModel.kt
@@ -1,4 +1,0 @@
-package com.example.android.vinylsappg21.viewmodels
-
-class CommentViewModel {
-}
