--- app/src/main/java/com/example/android/vinylsappg21/viewmodels/TrackViewModel.kt
@@ -1,4 +1,0 @@
-package com.example.android.vinylsappg21.viewmodels
-
-class TrackViewModel {
-}
